requirements.txt
strands-agents
anthropic
python-dotenv
rich
fastmcp
mcp